Rename Auth related files
