+++ server/src/web/controllers/AuthController.js
@@ -1,5 +1,5 @@
 import config from 'config';
-import * as AuthenticationService from 'services/AuthenticationService';
+import * as AuthService from 'services/AuthService';
 
 export function saveTarget(request, response, next) {
   request.session.targetUri = request.query.target;
@@ -7,7 +7,7 @@ export function saveTarget(request, response, next) {
 }
 
 export function createJWT(request, response, next) {
-  const token = AuthenticationService.createTokenForUser(request.user);
+  const token = AuthService.createTokenForUser(request.user);
 
   response.cookie('JWT', token, {
     maxAge: config.jwt.expiry * 60,
